+++ ReleaseNotes.md
@@ -1,3 +1,10 @@
+## What is New in VMware-NSXvCD-Module 1.1.0
+May, 2018
+
+No major changes.
+
+- Added 'NsxVcd' Prefix to  'Invoke-APICall' and 'New-APIConnection'
+
 ## What is New in VMware-NSXvCD-Module 1.0.0
 May, 2018